switch to free unicode fonts
This lets us do things like have bullet points for items in our
work done list, and lets people with unicode chars in their names
generate invoices.

Updated README to reflect the use of these specific fonts, how
to change fonts, and how to update the paths.

We really ought to have the font names and paths in the template
file, I suppose.