fine fine, I moved the font paths into the template
This code is untested, but i theory you can now specify new fonts
for sans/serif along with the paths to their ttf files, by
changing the appropriate settings in the yaml template.

Note that this would likely require tweaks to cell positioning
directly in the python script, so it's perhaps not that useful.

Updated README accordingly.