add optional border around displayed lists
Restructured the list displays so that header can be included,
multiple columns can be handled, border can be drawn, with no special
code needed in the caller

Also collected all the set_x(8) and made the left margin an attribute
which gets used instead, oof!

There's still a lot of hardcoded sizes in here but it's better than
it was.