remove BILLABLE_WIDTHS constant
We calculate withs of bordered tables from header lengths and page width

This now requires an extra class page_constant attribute, which I don't
love, but it's much better than having a bunch of numbers hardcoded in
there.

With the current setup this results in the subtotals, tax and total
column placed too far to the left. We need to make that table placement
independent of the billables table, to fix it.