replace all the constant text strings with a module constant
If someone wants to translate them or modify them or extend them,
they can go to one place to get it done.

Also changed "Week of:" to "Dates" since there is already
"Week of" in the txt of the week... ugh. another fixed string.
Well it can get dealt with some other time.